NEW-22: WIP - Improved structure of access analysis. Added more platforms and improved general structure

+++ include/astro/platforms/space/Plane.hpp
@@ -19,11 +19,11 @@ public:
     Plane() = default;
     Plane(std::vector<Spacecraft> satellites);
     ~Plane() = default;
-    
+
     void add_spacecraft(const Spacecraft& spacecraft);
 
     const std::vector<Spacecraft>& get_all_spacecraft() const;
-    
+
     const Spacecraft& get_spacecraft(const int& spacecraftId) const;
 
     const size_t size() const;
@@ -32,10 +32,10 @@ public:
 
 private:
 
-    int id;
+    size_t id;
     OrbitalElements elements;
     std::vector<Spacecraft> satellites;
-    
+
     bool strict;
 
     void generate_id_hash();
