+++ include/astro/platforms/space/Shell.hpp
@@ -20,25 +20,25 @@ public:
     Shell(std::vector<Plane> planes);
     Shell(std::vector<Spacecraft> satellites);
     ~Shell() = default;
-    
+
     const size_t size() const;
     const size_t n_planes() const;
 
     void add_plane(const Plane& plane);
     void add_spacecraft(const Spacecraft& spacecraft, const int& planeId);
     void add_spacecraft(const Spacecraft& spacecraft);
-    
+
     const std::vector<Plane>& get_all_planes() const;
     const std::vector<Spacecraft> get_all_spacecraft() const;
-    
+
     const Plane& get_plane(const int& planeId) const;
     const Spacecraft& get_spacecraft(const int& spacecraftId) const;
-    
+
     void propagate(EquationsOfMotion& eom, Integrator& integrator, const Interval& interval = Integrator::defaultInterval);
 
 private:
 
-    int id;
+    size_t id;
     std::string name;
     std::vector<Plane> planes;
